LayoutTests/sputnik/Conformance/12_Statement/12.9_The_return_Statement/S12.9_A1_T5-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Return statements are only valid inside functions
S12.9_A1_T5

PASS Expected parsing failure

TEST COMPLETE
